platformio tool complains about invalid json file
When I try to add this json file to platformio I get a cryptic error. The only possible problem I can see is the "dangling" comma. Please excuse me if you received this request before. I thought I created one but could not find it anymore, so I opened another one and hope to see it as a pull request.
Cheers, Dani